coffeelog/requirements.txt
click==7.0
gspread==3.1.0
oauth2client==4.1.3
google-cloud-texttospeech==0.5.0
